add preview monitor and split out some preview methods into the settings class
