+++ src/GraphQL/Auth.php
@@ -3,13 +3,13 @@
 namespace WPGatsby\GraphQL;
 
 use \Firebase\JWT\JWT;
-use \WPGatsby\Admin\Preview;
+use \WPGatsby\Admin\Settings;
 
 class Auth {
 	static function get_token() {
 		$site_url    = get_bloginfo( 'url' );
-		$preview_url = Preview::get_gatsby_preview_instance_url();
-		$secret      = Preview::get_setting( 'preview_jwt_secret' );
+		$preview_url = Settings::get_gatsby_preview_instance_url();
+		$secret      = Settings::get_setting( 'preview_jwt_secret' );
 		$now         = time();
 		$expiry      = $now + 3600;
 		$user_id     = get_current_user_id();
